visualization/requirements.txt
fastapi
uvicorn
jinja2
pydantic
python-multipart
requests
pyyaml
